# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateSecretRequest:
    properties:
      secret:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: My Personal API
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Syntaqx Personal API
  version: "1.0"
paths:
...
  /secrets:
    post:
      consumes:
      - application/json
      description: Create a secret
      parameters:
      - description: Secret
        in: body
        name: secret
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a secret
      tags:
      - secrets
  /secrets/{secretId}:
    get:
      consumes:
      - application/json
      description: Retrieve a secret
      parameters:
      - description: Secret ID
        in: path
        name: secretId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve a secret
      tags:
      - secrets
...
swagger: "2.0"